:truck: Move api-commons utility to a more related place
